[0ms][6MB] [Working solution] Fixed permutations function
Fixed the issue of floating point numbers which were only added to accommodate the possible floating point values.
On careful inspection realized that there shouldn't be a need of float points.
Fixed it using an auxiliary array which stores the dominator which are not factor of the current value of permutation.
Divide only when it's divisible.